requirements.txt
beautifulsoup4>=4.4.1
Pillow>=2.6.1
PyYAML>=3.11
cssselect>=0.9.1
lxml>=3.3.5
nltk>=3.2
requests>=2.3.0
feedparser>=5.1.3
tldextract>=1.5.1
feedfinder2>=0.0.4
jieba3k>=0.35.1
python-dateutil>=2.4.0
Pattern==2.6
six==1.10.0
urllib3==1.15.1
wheel==0.24.0
-e git+https://github.com/news-ai/newspaper@95297299b0a4db17e32b4a2d67533742de0ed876#egg=newspaper-origin_python-2-head